moved location of nvcategory_util.hpp and wrapper_types.hpp
