+++ cpp/benchmarks/type_dispatcher/type_dispatcher_benchmark.cu
@@ -24,7 +24,7 @@
 #include <tests/utilities/cudf_test_fixtures.h>
 #include <tests/utilities/cudf_test_utils.cuh>
 #include <cudf/types.hpp>
-#include <cudf/utils/wrapper_types.hpp>
+#include <cudf/utils/legacy/wrapper_types.hpp>
 
 #include <table/legacy/device_table.cuh>
