+++ cpp/src/binaryop/compiled/binary_ops.cu
@@ -20,7 +20,7 @@
 #include "binary_ops.hpp"
 #include <utilities/cudf_utils.h>
 #include <cudf/utils/legacy/type_dispatcher.hpp>
-#include <cudf/utils/wrapper_types.hpp>
+#include <cudf/utils/legacy/wrapper_types.hpp>
 #include <cudf/binaryop.hpp>
 
 namespace cudf {
