+++ cpp/src/copying/gather.cu
@@ -21,7 +21,7 @@
 #include <cudf/utils/legacy/type_dispatcher.hpp>
 
 #include <cudf/legacy/table.hpp>
-#include <cudf/utils/nvcategory_util.hpp>
+#include <cudf/utils/legacy/nvcategory_util.hpp>
 
 #include <algorithm>
 
@@ -31,7 +31,7 @@
 #include <cub/cub.cuh>
 #include <nvstrings/NVCategory.h>
 #include <nvstrings/NVStrings.h>
-#include <cudf/utils/nvcategory_util.hpp>
+#include <cudf/utils/legacy/nvcategory_util.hpp>
 #include <utilities/column_utils.hpp>
 #include <utilities/cuda_utils.hpp>
 using bit_mask::bit_mask_t;
