+++ cpp/src/io/csv/type_conversion.cuh
@@ -18,7 +18,7 @@
 #define CONVERSION_FUNCTIONS_CUH
 
 #include "datetime_parser.cuh"
-#include <cudf/utils/wrapper_types.hpp>
+#include <cudf/utils/legacy/wrapper_types.hpp>
 #include <cuda_runtime_api.h>
 
 #include <utilities/trie.cuh>
