+++ cpp/src/merge/merge.cu
@@ -19,7 +19,7 @@
 #include <table/legacy/device_table.cuh>
 #include <table/legacy/device_table_row_operators.cuh>
 #include "bitmask/legacy/bit_mask.cuh"
-#include <cudf/utils/nvcategory_util.hpp>
+#include <cudf/utils/legacy/nvcategory_util.hpp>
 #include "rmm/thrust_rmm_allocator.h"
 #include "utilities/cuda_utils.hpp"
