+++ cpp/src/reductions/reduction_operators.cuh
@@ -19,7 +19,7 @@
 
 #include <cudf/cudf.h>
 #include <utilities/cudf_utils.h>
-#include <cudf/utils/wrapper_types.hpp>
+#include <cudf/utils/legacy/wrapper_types.hpp>
 #include <utilities/error_utils.hpp>
 #include <iterator/iterator.cuh>
