+++ cpp/src/sort/segmented_sorting.cu
@@ -3,7 +3,7 @@
 #include <utilities/cudf_utils.h>
 #include <utilities/error_utils.hpp>
 #include <cudf/utils/legacy/type_dispatcher.hpp>
-#include <cudf/utils/wrapper_types.hpp>
+#include <cudf/utils/legacy/wrapper_types.hpp>
 
 
 #include <cub/device/device_segmented_radix_sort.cuh>
