+++ cpp/src/string/nvcategory_util.cpp
@@ -1,6 +1,6 @@
 #include <utility>
 #include <utilities/column_utils.hpp>
-#include <cudf/utils/nvcategory_util.hpp>
+#include <cudf/utils/legacy/nvcategory_util.hpp>
 #include <cudf/replace.hpp>
 #include <cudf/types.hpp>
 #include <cudf/legacy/table.hpp>
