+++ cpp/src/transform/transform.cpp
@@ -21,7 +21,7 @@
 #include <bitmask/legacy/bitmask_ops.hpp>
 #include <bitmask/legacy/legacy_bitmask.hpp>
 #include <cudf/copying.hpp>
-#include <cudf/utils/nvcategory_util.hpp>
+#include <cudf/utils/legacy/nvcategory_util.hpp>
 #include <utilities/error_utils.hpp>
 #include "jit/core/launcher.h"
