+++ cpp/src/unary/cast_ops.cu
@@ -25,7 +25,7 @@
 #include <cudf/copying.hpp>
 
 #include <rmm/thrust_rmm_allocator.h>
-#include <cudf/utils/wrapper_types.hpp>
+#include <cudf/utils/legacy/wrapper_types.hpp>
 #include <cudf/utils/legacy/type_dispatcher.hpp>
 
 #include <thrust/copy.h>
