+++ cpp/src/utilities/device_atomics.cuh
@@ -34,7 +34,7 @@
 
 #include <cudf/cudf.h>
 #include <utilities/cudf_utils.h>
-#include <cudf/utils/wrapper_types.hpp>
+#include <cudf/utils/legacy/wrapper_types.hpp>
 #include <utilities/error_utils.hpp>
 #include <utilities/device_operators.cuh>
