+++ cpp/tests/device_atomics/device_atomics_test.cu
@@ -16,7 +16,7 @@
 
 #include <tests/utilities/column_wrapper.cuh>
 #include <tests/utilities/cudf_test_fixtures.h>
-#include <cudf/utils/wrapper_types.hpp>
+#include <cudf/utils/legacy/wrapper_types.hpp>
 #include <utilities/device_atomics.cuh>
 
 #include <gmock/gmock.h>
