+++ cpp/tests/merge/merge-test.cu
@@ -13,7 +13,7 @@
 #include <rmm/thrust_rmm_allocator.h>
 #include <cudf/legacy/table.hpp>
 
-#include <cudf/utils/nvcategory_util.hpp>
+#include <cudf/utils/legacy/nvcategory_util.hpp>
 #include "tests/utilities/column_wrapper.cuh"
 #include "tests/utilities/column_wrapper_factory.hpp"
 #include "tests/utilities/cudf_test_fixtures.h"
