+++ cpp/tests/orderby/orderby_tests.cu
@@ -24,7 +24,7 @@
 #include <tests/utilities/tuple_vectors.h>
 
 #include <utilities/bit_util.cuh>
-#include <cudf/utils/wrapper_types.hpp>
+#include <cudf/utils/legacy/wrapper_types.hpp>
 #include <bitmask/legacy/legacy_bitmask.hpp>
 #include <cudf/cudf.h>
