+++ cpp/tests/types/types_test.cpp
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-#include <cudf/utils/wrapper_types.hpp>
+#include <cudf/utils/legacy/wrapper_types.hpp>
 #include <cudf/cudf.h>
 
 #include <gtest/gtest.h>
