+++ cpp/tests/utilities_tests/column_wrapper_tests.cu
@@ -17,7 +17,7 @@
 #include <tests/utilities/cudf_test_fixtures.h>
 #include <tests/utilities/column_wrapper.cuh>
 #include <cudf/utils/legacy/type_dispatcher.hpp>
-#include <cudf/utils/wrapper_types.hpp>
+#include <cudf/utils/legacy/wrapper_types.hpp>
 
 #include <gmock/gmock.h>
 #include <gtest/gtest.h>
